Integrated into SceneAnalyzer
