+++ Lorem Ipsum XR Translator/Assets/Scripts/ObjectDetection/AzureObjDetectionResponse.cs
@@ -26,23 +26,6 @@ namespace ObjectDetection
     }
 
     /// <summary>
-    /// Detected Object contains a rectangle location, object name, and confidence level
-    /// </summary>
-    public class DetectedObject
-    {
-        public Rectangle rectangle;
-        public string objectName;
-        public float confidence;
-
-        public DetectedObject(JObject objJson)
-        {
-            rectangle = new Rectangle((JObject)objJson["rectangle"]);
-            objectName = (string)objJson["object"];
-            confidence = (float)objJson["confidence"];
-        }
-    }
-
-    /// <summary>
     /// Metadata associated with image detection
     /// </summary>
     public class Metadata
