+++ Lorem Ipsum XR Translator/Assets/Scripts/ObjectDetection/DetectedObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a3e1048e5f29d44ebaa11150c819d7f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
