+++ Lorem Ipsum XR Translator/Assets/Scripts/ObjectDetection/IObjectDetectorClient.cs
@@ -10,7 +10,7 @@ namespace ObjectDetection
     /// </summary>
     public interface IObjectDetectorClient
     {
-        IEnumerator DetectObjects(string modelPath, string imagePath, Action<List<DetectedObject>> callback);
-        IEnumerator DetectObjects(string modelPath, Texture2D image, Action<List<DetectedObject>> callback);
+        IEnumerator DetectObjects(string imagePath, Action<List<DetectedObject>> callback);
+        IEnumerator DetectObjects(Texture2D image, Action<List<DetectedObject>> callback);
     }
 }
